+++ controllers/index.js
@@ -3,7 +3,7 @@ const router = require('express').Router();
 const apiRoutes = require('./api');
 const homeRoutes = require('./homeRoutes');
 
-router.use('/', homeRoutes);
-router.use('/api', apiRoutes);
+router.use('/', homeRoutes);  //front end
+router.use('/api', apiRoutes); // back end
 
 module.exports = router;
