+++ public/css/style.css
@@ -133,9 +133,9 @@ img {
 
 
 /* abigail's section for workout page */
-h1{
+/* h1{
   font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif;
-}
+} */
 
 image-wrapper{
   max-width: 100%;
