+++ views/homepage.handlebars
@@ -1,4 +1,4 @@
-{{ {{#each users as |user|}}
+{{#each users as |user|}}
   <div class="row align-center mb-5">
     <div class="col-md-6">
       <h2>{{{get_emoji}}} {{user.name}}</h2>
@@ -7,4 +7,4 @@
       <h3><a href="mailto:{{user.email}}">{{user.email}}</a></h3>
     </div>
   </div>
-{{/each}} }}
+{{/each}}
